pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aviation-eval-harness"
version = "0.1.0"
description = "Research-grade agent evaluation harness for aviation safety"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "anthropic>=0.39.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "sentence-transformers>=2.2",
    "matplotlib>=3.7",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
run-eval = "runners.run_eval:cli"

[tool.hatch.build.targets.wheel]
packages = ["evals", "models", "graders", "runners", "analysis", "datasets"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100
